requirements.txt
pyrogram
py-tgcalls==0.9.7
spotipy
aiohttp
requests
beautifulsoup4
ffmpeg-python
hachoir
pillow==10.0.1
pymongo
motor
python-dotenv
pyyaml
tgcrypto
unidecode
dnspython
aiofiles
psutil
httpx[http2]